opendal_service_gcs

Struct Gcs

pub struct Gcs { /* private fields */ }

Google Cloud Storage services support.

Capabilities

Depending on its configuration and the backing system, this service can expose:

create_dir

stat

read

write

delete

list

copy

rename

presign

Inspect the effective capability set with [opendal_core::Operator::info] and [opendal_core::OperatorInfo::capability] after building an operator.

Configuration

Use crate::GcsConfig for serializable configuration and this builder’s methods for direct construction. The field and method documentation defines accepted values, defaults, and environment interaction.

Options to authenticate to GCS

OpenDAL supports the following authentication options:

Provide a base64-ed JSON key string with credential

Provide a JSON key file at explicit path with credential_path

Provide a JSON key file at implicit path

GcsBackend will attempt to load Service Account key from ADC well-known places.

Fetch access token from VM metadata

Only works when running inside Google Cloud.

If a non-default Service Account name is required, set with service_account. Otherwise, nothing need to be set.

A custom credential provider chain via GcsBuilder.credential_provider_chain()

Notes:

When a Service Account key is provided, it will be used to create access tokens (VM metadata will not be used).

Explicit Service Account key, in json or path, always take precedence over ADC-defined key paths.

Due to limitation in GCS, a private key is required to create Pre-signed URL. Currently, OpenDAL only supports Service Account key.

Example

Via Builder

use opendal_core::Operator;
use opendal_core::Result;
use opendal_service_gcs::Gcs;

#[tokio::main]
async fn main() -> Result<()> {
    // create backend builder
    let mut builder = Gcs::default()
       // set the storage bucket for OpenDAL
       .bucket("test")
       // set the working directory root for GCS
       // all operations will happen within it
       .root("/path/to/dir")
       // set the credentials with service account
       .credential("service account JSON in base64")
       // set the predefined ACL for GCS
       .predefined_acl("publicRead")
       // set the default storage class for GCS
       .default_storage_class("STANDARD");

    let op: Operator = Operator::new(builder)?;
    Ok(())
}

Implementations

impl GcsBuilder

pub fn root(self, root: &str) -> Self

set the working directory root of backend

pub fn bucket(self, bucket: &str) -> Self

set the container’s name

pub fn scope(self, scope: &str) -> Self

set the GCS service scope

If not set, we will use https://www.googleapis.com/auth/devstorage.read_write.

Valid scope examples

read-only: https://www.googleapis.com/auth/devstorage.read_only

read-write: https://www.googleapis.com/auth/devstorage.read_write

full-control: https://www.googleapis.com/auth/devstorage.full_control

Reference: Cloud Storage authentication

pub fn service_account(self, service_account: &str) -> Self

Set the GCS service account.

service account will be used for fetch token from vm metadata. If not set, we will try to fetch with default service account.

pub fn endpoint(self, endpoint: &str) -> Self

set the endpoint GCS service uses

pub fn credential(self, credential: &str) -> Self

set the base64 hashed credentials string used for OAuth2 authentication.

this method allows to specify the credentials directly as a base64 hashed string. alternatively, you can use credential_path() to provide the local path to a credentials file. we will use one of credential and credential_path to complete the OAuth2 authentication.

Reference: Google Cloud Storage Authentication.

pub fn credential_path(self, path: &str) -> Self

set the local path to credentials file which is used for OAuth2 authentication.

credentials file contains the original credentials that have not been base64 hashed. we will use one of credential and credential_path to complete the OAuth2 authentication.

Reference: Google Cloud Storage Authentication.

pub fn credential_provider( self, provider: impl ProvideCredential<Credential = Credential> + 'static, ) -> Self

Specify a customized credential provider used by this service.

This provider will be pushed to the front of credential chain.

pub fn credential_provider_chain( self, chain: ProvideCredentialChain<Credential>, ) -> Self

Specify a customized credential provider chain used by this service.

This chain will be pushed to the front of default chain.

pub fn token(self, token: String) -> Self

Provide the OAuth2 token to use.

pub fn disable_vm_metadata(self) -> Self

Disable attempting to load credentials from the GCE metadata server.

pub fn disable_config_load(self) -> Self

Disable loading configuration from the environment.

pub fn predefined_acl(self, acl: &str) -> Self

Set the predefined acl for GCS.

Available values are:

authenticatedRead

bucketOwnerFullControl

bucketOwnerRead

private

projectPrivate

publicRead

pub fn default_storage_class(self, class: &str) -> Self

Set the default storage class for GCS.

Available values are:

STANDARD

NEARLINE

COLDLINE

ARCHIVE

pub fn skip_signature(self) -> Self

Skip signature will skip loading credentials and signing requests.

This is typically used for buckets which are open to the public or GCS storage emulators.

pub fn allow_anonymous(self) -> Self

👎Deprecated since 0.57.0:
Please use skip_signature instead of allow_anonymous

Allow anonymous requests.