opendal_service_gcs

Struct GcsConfig

#[non_exhaustive]
pub struct GcsConfig {

    pub root: Option<String>,
    pub bucket: String,
    pub endpoint: Option<String>,
    pub scope: Option<String>,
    pub service_account: Option<String>,
    pub credential: Option<String>,
    pub credential_path: Option<String>,
    pub predefined_acl: Option<String>,
    pub default_storage_class: Option<String>,
    pub skip_signature: bool,
    pub allow_anonymous: bool,
    pub disable_vm_metadata: bool,
    pub disable_config_load: bool,
    pub token: Option<String>,

}

Google Cloud Storage services support.

Fields (Non-exhaustive)

Non-exhaustive structs could have additional fields added in future. Therefore, non-exhaustive structs cannot be constructed in external crates using the traditional Struct { .. } syntax; cannot be matched against without a wildcard ..; and struct update syntax will not work.

root: Option<String>

root URI, all operations happens under root

bucket: String

bucket name

endpoint: Option<String>

endpoint URI of GCS service, default is https://storage.googleapis.com

scope: Option<String>

Scope for gcs.

service_account: Option<String>

Service Account for gcs.

credential: Option<String>

Credentials string for GCS service OAuth2 authentication.

credential_path: Option<String>

Local path to credentials file for GCS service OAuth2 authentication.

predefined_acl: Option<String>

The predefined acl for GCS.

default_storage_class: Option<String>

The default storage class used by gcs.

skip_signature: bool

Skip signature will skip loading credentials and signing requests.

allow_anonymous: bool
👎Deprecated since 0.57.0:
Please use skip_signature instead of allow_anonymous

Allow opendal to send requests without signing when credentials are not loaded.

disable_vm_metadata: bool

Disable attempting to load credentials from the GCE metadata server when running within Google Cloud.

disable_config_load: bool

Disable loading configuration from the environment.

token: Option<String>

A Google Cloud OAuth2 token.

Takes precedence over credential and credential_path.